App directory path feature provided for custom app.

+++ katana/primary_process.py
@@ -98,28 +98,33 @@ def install_custom_app(app, app_url):
     if not (os.path.exists(wapps_dir_path)):
         os.mkdir(wapps_dir)
     app_url= app_url.split(" ")
-    if len(app_url) == 3:
-        repo_url = app_url[0]
-        user_branch = app_url[2]
-    else:
-        repo_url = app_url[0]
-        user_branch = 'master'
     directory = app
-    tempdir = os.path.join(BASE_DIR, directory)
-    if (os.path.exists(tempdir)):
+    if (app_url[0].startswith("http")):
+        if len(app_url) == 3:
+            repo_url = app_url[0]
+            user_branch = app_url[2]
+        else:
+            repo_url = app_url[0]
+            user_branch = 'master'
+        tempdir = os.path.join(BASE_DIR, directory)
+        if (os.path.exists(tempdir)):
+            shutil.rmtree(tempdir)
+        os.mkdir(tempdir)
+        Repo.clone_from(repo_url, tempdir, branch= user_branch)
+        source = os.path.join(tempdir, directory)
+        destination = os.path.join(wapps_dir_path, directory)
+        if (os.path.exists(destination)):
+            shutil.rmtree(destination)
+        shutil.move(source, destination)
         shutil.rmtree(tempdir)
-    os.mkdir(tempdir)
-    Repo.clone_from(repo_url, tempdir, branch= user_branch)
-    source = os.path.join(tempdir, directory)
-    destination = os.path.join(wapps_dir_path, directory)
-    if (os.path.exists(destination)):
-        shutil.rmtree(destination)
-    shutil.move(source, destination)
-    shutil.rmtree(tempdir)
-    configure_settings_file_custom_app(app)
-    # print("DONE !\n")
+    else:
+        destination = os.path.join(wapps_dir_path, directory)
+        app_path = app_url[0]
+        if(os.path.isdir(app_path)):
+            shutil.copytree(app_path,destination)
     configure_urls_file_custom(app, "wapps")
-    # print("DONE !\n\n")
+    configure_settings_file_custom_app(app)
+
 
 def remove_cust_app_source(uapp,  category):
     if category == "wapps":
